fix: fix the flicker when liked list is updated
