+++ mobile/app/(tabs)/index.jsx
@@ -16,10 +16,11 @@ import { useEffect, useRef, useState } from "react";
 import styles from "../../assets/styles/home.styles";
 import { API_URL } from "../../constants/api";
 import { Ionicons } from "@expo/vector-icons";
-import { formatPublishDate } from "../../lib/utils";
 import COLORS from "../../constants/colors";
 import Loader from "../../components/Loader";
 import { sleep } from "../../lib/utils";
+import { useCallback } from "react";
+import ReviewCard from "../../components/ReviewCard";
 
 export default function Home() {
   const { token } = useAuthStore();
@@ -42,7 +43,7 @@ export default function Home() {
     try {
       refresh ? setRefreshing(true) : pageNum === 1 && setLoading(true);
 
-      const res = await fetch(`${API_URL}/reviews?page=${pageNum}&limit=4`, {
+      const res = await fetch(`${API_URL}/reviews?page=${pageNum}&limit=2`, {
         headers: { Authorization: `Bearer ${token}` },
       });
 
@@ -123,111 +124,16 @@ export default function Home() {
     }
   };
 
-  /* ---------- Item component ---------- */
-  const ReviewCard = ({ item }) => {
-    const isLiked = likedIds.has(item._id);
-    const scaleAnim = useRef(new Animated.Value(0)).current;
-    const lastTap = useRef(null);
-
-    const animateHeart = () => {
-      scaleAnim.setValue(0.3);
-      Animated.sequence([
-        Animated.spring(scaleAnim, { toValue: 1, useNativeDriver: true }),
-        Animated.timing(scaleAnim, {
-          toValue: 0,
-          duration: 400,
-          useNativeDriver: true,
-        }),
-      ]).start();
-    };
-
-    const handleDoubleTap = () => {
-      const now = Date.now();
-
-      if (lastTap.current && now - lastTap.current < 300) {
-        animateHeart();
-        if (!isLiked) {
-          toggleFavourite(item._id);
-        }
-      } else {
-        lastTap.current = now;
-      }
-    };
-
-    return (
-      <View style={styles.bookCard}>
-        <View style={styles.bookHeader}>
-          <View style={styles.row}>
-            <View style={styles.userInfo}>
-              <Image
-                source={{ uri: item.user.profileImage }}
-                style={styles.avatar}
-              />
-              <Text style={styles.username}>{item.user.username}</Text>
-              {/* Heart button */}
-              <TouchableOpacity
-                onPress={() => toggleFavourite(item._id)}
-                style={styles.heartAbs}
-              >
-                <Ionicons
-                  name={isLiked ? "heart" : "heart-outline"}
-                  size={24}
-                  color={isLiked ? COLORS.primary : COLORS.textSecondary}
-                />
-              </TouchableOpacity>
-            </View>
-          </View>
-        </View>
-
-        {/* Image with double-tap */}
-        <TouchableWithoutFeedback onPress={handleDoubleTap}>
-          <View style={styles.bookImageContainer}>
-            <Image
-              source={item.image}
-              style={styles.bookImage}
-              contentFit="cover"
-            />
-
-            {/* big heart animation */}
-            <Animated.View
-              style={[
-                styles.bigHeart,
-                { transform: [{ scale: scaleAnim }], opacity: scaleAnim },
-              ]}
-            >
-              <Ionicons name="heart" size={120} color="rgba(255,255,255,0.9)" />
-            </Animated.View>
-          </View>
-        </TouchableWithoutFeedback>
-
-        {/* Details */}
-
-        <View style={styles.bookDetails}>
-          <Text style={styles.bookTitle}>{item.title}</Text>
-
-          <View style={styles.ratingContainer}>
-            {renderRatingStars(item.rating)}
-          </View>
-          <Text style={styles.caption}>{item.caption}</Text>
-          <Text style={styles.date}>
-            Shared on {formatPublishDate(item.createdAt)}
-          </Text>
-        </View>
-      </View>
-    );
-  };
-
-  /** RATING STARS *********************************************************/
-  const renderRatingStars = (rating) =>
-    Array.from({ length: 5 }, (_, i) => (
-      <Ionicons
-        key={i + 1}
-        name={i + 1 <= rating ? "star" : "star-outline"}
-        size={16}
-        color={i + 1 <= rating ? "#f4b400" : COLORS.textSecondary}
-        style={{ marginRight: 2 }}
+  const renderItem = useCallback(
+    ({ item }) => (
+      <ReviewCard
+        item={item}
+        isLiked={likedIds.has(item._id)}
+        toggleFavourite={toggleFavourite}
       />
-    ));
+    ),
+    [likedIds],
+  );
 
   if (loading) return <Loader />;
 
@@ -236,7 +142,7 @@ export default function Home() {
     <View style={styles.container}>
       <FlatList
         data={reviews}
-        renderItem={({ item }) => <ReviewCard item={item} />}
+        renderItem={renderItem}
         keyExtractor={(item) => item._id}
         contentContainerStyle={styles.listContainer}
         showsVerticalScrollIndicator={false}
